server/requirements.txt
# Python dependencies for Raspberry Pi scripts
# Install with: pip install -r requirements.txt

python-dotenv==1.0.0
mysql-connector-python==8.2.0
RPi.GPIO==0.7.1
mfrc522==0.0.7
RPLCD==1.3.1
